requirements.txt
flask
flask-cors
python-dotenv
pyjwt
reportlab
gunicorn
